Move customgeostylerproperties.py and avoid no qgis error
