--- .github/workflows/notify.yaml
@@ -1,19 +1,0 @@
-name: 'Notify camptocamp/lyrx2sld on updates'
-
-on:
-  push:
-    branches:
-      - master
-
-jobs:
-  notify:
-    name: 'Notify lyrx2sld'
-    runs-on: ubuntu-latest
-    steps:
-      - name: Notify lyrx2sld
-        run: >
-          curl --request POST --header "Content-Type: application/json"
-          --header 'Accept: application/vnd.github+json'
-          --header "Authorization: token ${{ secrets.GOPASS_CI_GITHUB_TOKEN }}"
-          https://api.github.com/repos/camptocamp/lyrx2sld/dispatches
-          --data '{"event_type": "bridge-style_updated"}'
