+++ bridgestyle/geostyler/custom_properties.py
